Reorganized and commented code
